(PC-12840) feat(accessibility): Refactor TabBar to reuse it for web tap bar
